+++ packages/ui/src/types/authenticator/authUser.ts
@@ -1,0 +1,19 @@
+import { CognitoUser } from 'amazon-cognito-identity-js';
+
+export enum AuthChallengeNames {
+  SMS_MFA = 'SMS_MFA',
+  SOFTWARE_TOKEN_MFA = 'SOFTWARE_TOKEN_MFA',
+  NEW_PASSWORD_REQUIRED = 'NEW_PASSWORD_REQUIRED',
+  RESET_REQUIRED = 'RESET_REQUIRED',
+  MFA_SETUP = 'MFA_SETUP',
+}
+
+export interface CognitoAttributes {
+  email: string;
+  phone_number: string;
+  [key: string]: string;
+}
+export interface CognitoUserAmplify extends CognitoUser {
+  username?: string;
+  attributes?: CognitoAttributes;
+}
