+++ packages/ui/src/types/authenticator/stateMachine/authMachine.ts
@@ -1,0 +1,11 @@
+/**
+ * This files provides types that describe general shape of
+ * authenticator machine and its intepreter.
+ */
+import { Interpreter } from 'xstate';
+import { AuthContext } from './authContext';
+import { AuthEvent } from './authEvent';
+
+export type AuthInterpreter = Interpreter<AuthContext, any, AuthEvent>;
+
+export type AuthMachineSend = AuthInterpreter['send'];
